requirements.txt
pyaudio==0.2.14
pyttsx3==2.99
SpeechRecognition==3.14.5
pywin32==311
requests>=2.28.0
python-dotenv>=1.0.0
